Fix read() function err checking
The read function is failing tests with the inability to spawn a process with the error process not running. Upon investigation, I determined that the fact the read function had an || instead of a && caused a "pty: bad file descriptor" error and would not allow for further functions to execute and read from the buffer.