fix SpawnFake session ending
This was detected by using Tee and checking if the Tee `Close` function was
called. It was not ... adding the `srv.Close()` call made `waitForSession`
end (the <-endch whould no lomger block).
But the Tee `Close` function was called twice, because the batcher `srv` is
also using Tee while it should not.

Signed-off-by: Julien Viard de Galbert <julien.viard-de-galbert@itrenew.com>